+++ changelog.md
@@ -1,17 +1,22 @@
 # Version 1.88.2 (2025-10-06)
 
-- Bump aiohomematic to 2025.10.3
-  - Add Keyword-only method linter (mypy-style)
+## What's Changed
+- Bump aiohomematic to 2025.10.4
   - API cleanup: ensure that kw arguments are passed to the underlying function
+  - Add Keyword-only method linter (mypy-style)
+  - Add option to delay new device creation
+  - Add source of device creation to callback
 - Add unknown_pong_mismatch notification
+- Add option to config flow to delay new device creation
 
 # Version 1.88.1 (2025-10-01)
 
+## What's Changed
 - Bump aiohomematic to 2025.10.1
   - Add option to cover entities that the current default behaviour can be disabled:
       - The default behaviour is, that the primary cover entity of a group uses the level of the state channel and no its own level to display a correct level.
       - Only HM experts should disable this option, that like to control all three writeable channels of a cover group. 
-  - Add option to config_flow for cover entities
+  - Add option to config flow for cover entities
 
 # Version 1.88.0 (2025-10-01)